+++ packages/extension/src/preferences/preferences.tsx
@@ -1,0 +1,6 @@
+import ReactDOM from 'react-dom/client';
+import PreferencesForm from './PreferencesForm';
+
+ReactDOM.createRoot(document.querySelector('#root')!).render(
+    <PreferencesForm />
+);